Add support for local socket connections
This patch adds support for connecting to Libvirt via local socket.
In many cases users are running WebVirtMgr on the same system as
Libvirt, such as on an unRAID deployment. Connecting via local
socket works with "out of the box" Libvirt deployments and is easier
to configure for most users.